Use a single instance of Serializer per Cluster
Serializer contains information of the protocol version and type serializers (codecs) that are shared across instances created from the Cluster.

+++ src/Cassandra/FrameHeader.cs
@@ -88,7 +88,7 @@ namespace Cassandra
         /// Gets the size of the protocol header, depending on the version of the protocol
         /// </summary>
         /// <param name="version">Version of the protocol used</param>
-        public static byte GetSize(byte version)
+        public static byte GetSize(ushort version)
         {
             if (version >= 3)
             {
@@ -100,7 +100,7 @@ namespace Cassandra
         /// <summary>
         /// Parses the first 8 or 9 bytes and returns a FrameHeader
         /// </summary>
-        public static FrameHeader ParseResponseHeader(byte version, byte[] buffer, int offset)
+        public static FrameHeader ParseResponseHeader(ushort version, byte[] buffer, int offset)
         {
             var header = new FrameHeader()
             {
@@ -125,7 +125,7 @@ namespace Cassandra
         /// <summary>
         /// Parses the first 8 or 9 bytes from multiple buffers and returns a FrameHeader
         /// </summary>
-        public static FrameHeader ParseResponseHeader(byte version, byte[] buffer1, byte[] buffer2)
+        public static FrameHeader ParseResponseHeader(ushort version, byte[] buffer1, byte[] buffer2)
         {
             var headerBuffer = new byte[version < 3 ? 8 : 9];
             Buffer.BlockCopy(buffer1, 0, headerBuffer, 0, buffer1.Length);
